- write Odyssey in the middle - extract sidebar toggles in separate component - add dummy language chooser icon and position it in header
