+++ src/odyssey-web/src/App.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import LocalizationProvider from './contexts/LocalizationProvider';
-import Header from './components/Layout/Header';
+import Header from './components/header/Header';
 import SidebarLeft from './components/Layout/SidebarLeft';
 import SidebarRight from './components/Layout/SidebarRight';
 import DacInput from './components/DacInput';
