+++ src/odyssey-web/src/components/language-selector/LanguageSelector.scss
@@ -1,0 +1,11 @@
+.language-selector {
+    position: absolute;
+    right: 15px;
+    cursor: pointer;
+
+    button {
+        background: white;
+        border: none;
+        outline: none;
+    }
+}
