+++ src/odyssey-web/src/components/sidebar-toggle-button/SidebarToggleButton.scss
@@ -1,0 +1,68 @@
+@use '../../styles/variables' as variables;
+
+.hamburger,
+.hamburger:before,
+.hamburger:after {
+    cursor: pointer;
+    border-radius: 1px;
+    height: 3px;
+    width: variables.$hamburger-width-middle;
+    background: variables.$hamburger-bg;
+    display: block;
+    content: '';
+    transition: variables.$transition-appearance-fast;
+}
+
+.hamburger {
+    position: relative;
+}
+
+.hamburger:before {
+    top: -8px;
+    width: variables.$hamburger-width-top;
+    position: absolute;
+}
+
+.hamburger:after {
+    bottom: -8px;
+    width: variables.$hamburger-width-bottom;
+    position: absolute;
+}
+
+.sidebar-toggle:hover {
+
+    .hamburger,
+    .hamburger:before,
+    .hamburger:after {
+        background: variables.$primary;
+    }
+}
+
+.hamburger-right {
+
+    &,
+    &:before,
+    &:after {
+        right: 0;
+    }
+}
+
+.sidebar-toggle {
+    cursor: pointer;
+    width: 26px;
+    height: 26px;
+    display: flex;
+
+    &.right {
+        margin-left: variables.$spacer;
+    }
+
+    &.left {
+        margin-right: variables.$spacer;
+    }
+}
+
+button.sidebar-toggle {
+    background: none;
+    border: none;
+}
