+++ src/odyssey-web/src/styles/App.scss
@@ -10,21 +10,15 @@
 
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap');
 
-
-.header {
-  display: flex;
-  align-items: center;
-  padding-left: variables.$header-padding-x;
-  padding-right: variables.$header-padding-x;
-  padding-top: variables.$header-padding-y;
-  padding-bottom: variables.$header-padding-y;
-
-}
-
-button.sidebar-toggle {
-  background: none;
-  border: none;
-}
+//REMOVE
+// .header {
+//   display: flex;
+//   align-items: center;
+//   padding-left: variables.$header-padding-x;
+//   padding-right: variables.$header-padding-x;
+//   padding-top: variables.$header-padding-y;
+//   padding-bottom: variables.$header-padding-y;
+// }
 
 
 #root {
