+++ src/odyssey-web/src/styles/_variables.scss
@@ -67,8 +67,9 @@ $border-radius: .2rem !default;
 $border-radius-lg: .3rem !default;
 
 // Header
-$header-padding-y: $spacer*0.6 !default;
-$header-padding-x: $spacer*1.0 !default;
+//TODO REMOVE
+// $header-padding-y: $spacer*0.6 !default;
+// $header-padding-x: $spacer*1.0 !default;
 
 // Navbar
 // TODO REMOVE
